+++ packages/components/components/elvis-popover/src/react/elvia-popover.tsx
@@ -240,7 +240,9 @@ const Popover: FC<PopoverProps> = ({
     const triggerElementPositionPosition = popoverTriggerRef.current.getBoundingClientRect();
 
     // check for current fixed area
-    if (popoverFixedAreaRef.current != null) {
+    if (popoverFixedAreaRef.current === null || popoverContentRef.current === null) {
+      return;
+    } else {
       // define height and width to fixed area to match trigger element & and set top.
       popoverFixedAreaRef.current.style.top = triggerElementPositionPosition.y + 'px';
       popoverFixedAreaRef.current.style.height = triggerElementPositionPosition.height + 'px';
@@ -248,53 +250,50 @@ const Popover: FC<PopoverProps> = ({
     }
   };
 
-  const removeStylesOnClosed = () => {
-    // if popovervisibility false then return & remove any applied styles.
+  // place fixed area that covers trigger element and works as position anchor for content element
+  useEffect(() => {
+    // if false then return & remove any applied styles.
     if (!popoverVisibility) {
       if (popoverFixedAreaRef.current) {
         popoverFixedAreaRef.current.style.height = '0px';
         popoverFixedAreaRef.current.style.width = '0px';
       }
-      return false;
-    }
-  };
-
-  const isConflictTopOrBottom = () => {
-    // apply top and bottom properties if not enought space above or below popover
-    if (isConflictTop() && popoverContentRef.current && popoverTriggerRef.current) {
-      popoverContentRef.current.style.top = popoverTriggerRef.current.getBoundingClientRect().height + 'px';
-      popoverContentRef.current.style.bottom = 'auto';
-    }
-    if (
-      popoverClasscontainerRef.current?.classList.contains('ewc-popover--bottom') &&
-      popoverContentRef.current &&
-      popoverTriggerRef.current
-    ) {
-      popoverContentRef.current.style.top = popoverTriggerRef.current.getBoundingClientRect().height + 'px';
-    }
-  };
-
-  // place fixed area that covers trigger element and works as position anchor for content element
-  useEffect(() => {
-    if (!removeStylesOnClosed()) {
       return;
     }
+    // get triggerelement and check that it´s not null
+    const triggerElement = popoverTriggerRef.current;
+    if (triggerElement === null) {
+      return;
+    }
+    // get height of trigger element.
+    const triggerElementPositionHeight = triggerElement.getBoundingClientRect().height;
 
-    // Define size for current fixed container
-    if (popoverFixedAreaRef.current != null || popoverContentRef.current != null) {
+    // Check for and define size ffor current fixed container
+    if (popoverFixedAreaRef.current === null || popoverContentRef.current === null) {
+      return;
+    } else {
+      // define height and width to fixed area to match trigger element & and set top.
       defineFixedArea();
     }
-
-    // check if enought space on top or below the popover
-    isConflictTopOrBottom();
-
+    // apply top and bottom properties if not enought space above popover
+    if (isConflictTop()) {
+      popoverContentRef.current.style.top = triggerElementPositionHeight + 'px';
+      popoverContentRef.current.style.bottom = 'auto';
+    }
+    // apply top property if popver has bottom class
+    if (popoverClasscontainerRef.current?.classList.contains('ewc-popover--bottom')) {
+      if (popoverContentRef.current != null) {
+        popoverContentRef.current.style.top = triggerElementPositionHeight + 'px';
+      }
+    }
     // on scroll, reposition fixed area to current triggerelement position
     const UpdateContentPositionOnScroll = () => {
-      if (popoverFixedAreaRef.current && popoverTriggerRef.current) {
-        popoverFixedAreaRef.current.style.top = popoverTriggerRef.current.getBoundingClientRect().top + 'px';
+      if (popoverFixedAreaRef.current != null) {
+        popoverFixedAreaRef.current.style.top = triggerElement.getBoundingClientRect().top + 'px';
       }
     };
     document.addEventListener('scroll', UpdateContentPositionOnScroll, false);
+
     // Cleanup
     return () => {
       // Remove scroll listener
@@ -325,7 +324,7 @@ const Popover: FC<PopoverProps> = ({
         <div className="ewc-popover__backdrop" ref={popoverBackdropRef}></div>
 
         <div className="ewc-popover__fixed-content-area" ref={popoverFixedAreaRef}>
-          <div className="ewc-popover__contentContainer">
+          <div className="ewx-popover__contentContainer">
             <div className="ewc-popover__content" ref={popoverContentRef}>
               {hasCloseBtn == true && (
                 <div className="ewc-popover__close">