+++ packages/web/src/app/shared/component-documentation/component-structure/component-section/component-section.component.ts
@@ -11,8 +11,9 @@ export class ComponentSectionComponent {
   @Input() sectionTitle = '';
   @Input() propertiesClass = '';
   @Input() noAnimation = false;
+  @Input() figmaOnly = false;
 
-  constructor(private router: Router, private copyService: CopyToClipboardService) {}
+  constructor(private router: Router, private copyService: CopyToClipboardService) { }
 
   copyAnchor(): void {
     const anchorTitleElement = document.getElementById(this.sectionTitle);